Schedule email tasks and additionally provide recipients via CSV files
This commit does a few things:

- Add a HTTP endpoint to push tasks to the Redis queue. This endpoint is
  generic and can be used for more and different tasks in the future.
- Provide an interface for admins to add recipients. There is a textarea
  input that can be used to manually enter one recipient per line.
- Upload CSV files containing recipients. One or more CSV files can be
  uploaded and recipients are added cumulative. This allows a mix and
  match approach between uploading CSV files and editing email addresses
  in the browser.

I created an imaginary task for vote invitations. I believe this is
similar to what #136 will require. This part might require refactoring
when we get to it though.

I added tests for the HTTP endpoints but have no idea how to test the
task queueing or the actual email sending.