Allow only a single question for a festival
While artworks can have multiple questions festivals are only permitted
to have a single question. This commit enforces this constraint on the
database level by introducing a question type column. The API and the
client are updated to submit the type of the question along with the
question.

Unfortunately this change broke the vote and results unit tests. They
might require a bigger refactor to make them pass again.

closes #147