Initialize answers inline of the question view.
Answers can be initialized at two places now, inline of the answers
table on the question view and on the edit answer view, the same as
before. I kept the edit page intact since we also use that view to
delete answers.

In order to reduce code duplication I created a new `InitializeAnswer`
components whose purpose it is to manage the logic of initializing
answers. That component is used on the answers table as well as on the
edit answers view. It manages the logic and calls a render
prop (`children`) as a function to render the UI itself.

The table implementation is very limited and doesn't allow for inline
actions (and from what I can tell neither multiple actions on a table
row). I created special support for an `ACTION_ANSWER_INITILIAZE` table
action that is treated special inside the table. If we ever have to
support more inline actions we might want to refactor this approach.

Having multiple actions leads in this case to a rather ugly line break
of the action buttons. Shortening the `Deactivate` label by one
character would solve the problem. Otherwise we need to edit the styles
of the `ButtonOutline` component. I didn't want to do that for this case
since I'm not totally certain of the repercussions to the styling of the
rest of the page.