Create a question inline during festival creation.
This commit adds a question field to the festival new and edit form. It
creates or updates the question inline the festival page.

This works with the exception of deleting festivals, which does not
cascade and deletes the appropriate question as well, but only sets the
`festivalId` field to `NULL`.

Further improvements that can happen:

- Initialize questions inline of the festival page. Can I deactivate
  questions without deactivating festivals?
- Prevent deletion of questions in the questions panel, since both the
  festival and question have to stay intact. Currently if either of the
  two gest deleted the UI breaks.
- Move the relation between a festival and a question from the question
  to the festival. This encodes the 1:1 relationship between festivals
  and questions on the database layer.